+++ config/config.go
@@ -615,6 +615,7 @@ func getBootstrapConfig(v *viper.Viper, networkID uint32) (node.BootstrapConfig,
 		BootstrapMaxTimeGetAncestors:            v.GetDuration(BootstrapMaxTimeGetAncestorsKey),
 		BootstrapAncestorsMaxContainersSent:     int(v.GetUint(BootstrapAncestorsMaxContainersSentKey)),
 		BootstrapAncestorsMaxContainersReceived: int(v.GetUint(BootstrapAncestorsMaxContainersReceivedKey)),
+		SkipToSnowman:                           v.GetBool(SkipToSnowmanKey),
 	}
 
 	// TODO: Add a "BootstrappersKey" flag to more clearly enforce ID and IP