+++ config/keys.go
@@ -75,6 +75,7 @@ const (
 	StateSyncIDsKey                                      = "state-sync-ids"
 	BootstrapIPsKey                                      = "bootstrap-ips"
 	BootstrapIDsKey                                      = "bootstrap-ids"
+	SkipToSnowmanKey                                     = "skip-to-snowman"
 	StakingHostKey                                       = "staking-host"
 	StakingPortKey                                       = "staking-port"
 	StakingEphemeralCertEnabledKey                       = "staking-ephemeral-cert-enabled"